+++ src/main/java/team/mediasoft/wareshop/entity/dto/product/ProductDtoInfo.java
@@ -3,7 +3,7 @@ package team.mediasoft.wareshop.entity.dto.product;
 import lombok.Getter;
 import lombok.Setter;
 import team.mediasoft.wareshop.entity.enumeration.ProductCategory;
-import team.mediasoft.wareshop.exchanger.model.Currency;
+import team.mediasoft.wareshop.integration.currency.model.Currency;
 
 import java.math.BigDecimal;
