--- src/main/java/team/mediasoft/wareshop/exchanger/service/CurrencyServiceClient.java
@@ -1,7 +1,0 @@
-package team.mediasoft.wareshop.exchanger.service;
-
-import team.mediasoft.wareshop.exchanger.model.ExchangeRate;
-
-public interface CurrencyServiceClient {
-    ExchangeRate getExchangeRate();
-}
