+++ src/main/java/team/mediasoft/wareshop/integration/account/CustomerMockAccountServiceImpl.java
@@ -1,4 +1,4 @@
-package team.mediasoft.wareshop.businesslogic.service.account;
+package team.mediasoft.wareshop.integration.account;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Primary;
