+++ src/main/java/team/mediasoft/wareshop/integration/currency/service/CurrencyServiceClient.java
@@ -1,0 +1,7 @@
+package team.mediasoft.wareshop.integration.currency.service;
+
+import team.mediasoft.wareshop.integration.currency.model.ExchangeRate;
+
+public interface CurrencyServiceClient {
+    ExchangeRate getExchangeRate();
+}
