+++ src/main/java/team/mediasoft/wareshop/service/ProductService.java
@@ -9,9 +9,9 @@ import team.mediasoft.wareshop.entity.dto.product.ProductCreateEditDto;
 import team.mediasoft.wareshop.entity.dto.product.ProductReadDto;
 import team.mediasoft.wareshop.entity.dto.product.ProductUpdateDto;
 import team.mediasoft.wareshop.exception.ProductNotFoundException;
-import team.mediasoft.wareshop.exchanger.model.Currency;
-import team.mediasoft.wareshop.exchanger.model.CurrencyProvider;
-import team.mediasoft.wareshop.exchanger.service.ExchangeRateProvider;
+import team.mediasoft.wareshop.integration.currency.model.Currency;
+import team.mediasoft.wareshop.integration.currency.model.CurrencyProvider;
+import team.mediasoft.wareshop.integration.currency.service.ExchangeRateProvider;
 import team.mediasoft.wareshop.mapper.ProductMapper;
 import team.mediasoft.wareshop.search.criteria.SearchCriteria;
 import team.mediasoft.wareshop.search.specification.ProductSpecification;
